Implement correct display of nested rich texts
There's a lot of shenanigans with the functional components, as it is apparently the only way to return multiple nodes from a component. If rich-text were a functional component, it wouldn't need to deal with the container at all, but there's some problem with registration of recursive components outside of a template.